Project 3
This project is about using threads to be a basic web server instead of the web browser implementation of projects 1 and 2. See the ReadME.md for more details and be sure to unzip the testing.zip for the examples that the web server is able to generate. Requires WGET commands and connection to UMN VPN to work